move help output functions to help.go
